pyproject.toml
[project]
name = "rag-llm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "unstructured[md]==0.14.4",
    "loguru>=0.7.3",
    "python-dotenv==1.0.1",
    "langchain==0.2.2",
    "langchain-community==0.2.3",
    "langchain-openai==0.1.8",
    "chromadb==0.5.0",
    "openai==1.31.1",
    "tiktoken==0.7.0",
    "langchain-google-genai>=1.0.10",
]
